network_bytes metrics also need to account for previous connections
The network_bytes metrics were calculated only from metrics of the currently
active connections. After a connection has been disconnected, its so
far transmitted and received bytes were simply forgot. This lead to undefined
effects on the monitoring, since the metrics are assumed to be monotonically
increasing by design... which then confused Prometheus and finally myself :)

This fix adapts the solution in the NodeJS code to the Rust implementation. When
a peer leaves, its metrics are added to the accumulated past metrics in
connection_pool. When the overall metrics are then requested, these metrics are
added to the metrics of all currently connected peers to obtain the overall value.

The credit for the comment about why unwrap is safe to use in this patch goes
to @jeff. Thanks! :)